requirements.txt
numpy>=1.11.1
scipy>=0.18.1
python-dateutil>=2.4.0

# Style check and fix
autopep8 >= 1.2.4
    pycodestyle >=2.0.0
flake8 >= 3.0.4
    mccabe >= 0.5.0, < 0.6.0
    pyflakes >= 0.8.1, != 1.2.0, != 1.2.1, != 1.2.2, < 1.3.0
    configparser; python_version<"3.2"
    enum34; python_version<"3.4"
